refactor(eval): establish repository evaluation module
