+++ evaluation/__init__.py
@@ -1,0 +1,1 @@
+"""Repository-local evaluation tooling (not part of the installed apxinf package)."""
